etl/requirements.txt
# ETL Pipeline Dependencies
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
pyyaml==6.0.1
pymysql==1.1.0
sqlalchemy==2.0.23
pyarrow==14.0.1
duckdb==0.9.2
python-dotenv==1.0.0
tqdm==4.66.1

